// frontend/src/types/models/Language.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Language = {
	/**
	 * Resource Id.
	 */
	id: string;
	/**
	 * The language name.
	 */
	name: string;
	/**
	 * The language level.
	 */
	level: string;
	/**
	 * The language certificate url.
	 */
	certificateLink?: string;
};